:zap: first pass at respimgs
Using the data in the gallery YML, Jekyll spits out srcset/sizes
for images individually. The browser gets to pick the source it
wants, then we copy that to the <picture> background-image just
like the previous <span>-based galleries in order to preserve
the layout adjustments I have been making via CSS.

For now, this means the gallery has single-download responsive
images without having to go back and do a massive audit of all
the images I have created so far. \o/