:nut_and_bolt: 'mostly' category-based tags
This is good, not perfect. The tag page creator makes
an assumption about the category in order to provide
tags in two different directories.

The good part is that it will scale to a third category
if ever needed. The bad part is that you can't reuse
tags between two categories.